[Fix] for the BaseAttackController's params unknown problem, class devide
